+++ example/lib/src_zh/home.dart
@@ -100,7 +100,7 @@ class HomePageScaffold extends StatelessWidget {
                     color: Colors.black,
                     fontSize: 16.sp,
                   ),
-                  textScaleFactor: 1.0,
+                  textScaler: TextScaler.noScaling,
                 ),
                 Text(
                   '我的文字大小在设计稿上是16dp，会随着系统的文字缩放比例变化',