+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 5.9.3
+feat: support for disabling scaling
+
+# 5.9.2
+- Add ability to exclude widgets from from rebuild list
+- Depend on View instead of MediaQuery
+- update readme
+
+# 5.9.1
+- merge pr #550 #555 #556
+- update readme
+
 # 5.9.0
 - ScreenUtilInit won't rebuild the whole widget tree
 - Add `fontSizeResolver` to specify how font size should be scaled
@@ -20,9 +32,6 @@
 
 # 5.8.1
 - Limit flutter version > 3.10.0
-
-# 5.8.0
-- mirage flutter 3.10
 
 # 5.7.0
 - merge #464